'cumsegagg' in progress, paving the way to snapshotting.
